#cele-25 Compare workspace dialog
